refactor: moved a variety of image treatment functions to `cogs/utils/image_helpers.py`, rewrote `filter_image` to be used as a function, allowing image transformations to be run in parallel (previously could not be done due to pickling issue).
